# swagger.json
{
  "openapi": "3.0.0",
  "components":{
    "securitySchemes":{
      "ourbearer": {
        "type": "http",
        "scheme": "bearer",
        "in": "header",
        "name":"Authorization"
      }
    }
  },
  "security":[{
    "ourbearer":[]
  }]
    ,

  "info": {
    "title": "Your API",
    "description": "API documentation",
    "version": "1.0.0"
  },
  "paths": {
    "/get-employees": {
      "get": {
        "summary": "Get all employees",
        "description": "Returns a list of all users",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "current page",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "number of employees at a time",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],

        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "name": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/only-admin": {
      "get":{
        "summary": "Only admin can access",
        "description": "Admin can do multiple operations",
        "produces":["application/json"],
        "consumes":["application/json"],
        "responses": {
          "200":{
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "email": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "403":{
            "content":{
              "text/plain":{
                "description": "Not allowed to this resource"
              }
            }
          }
        }
      }
    },
    "/create-employee": {
      "post": {
        "summary": "Create list of employees",
        "requestBody": {
          "content": {
            "application/json": {
              "description": "request post object",
              "schema": {
                "$ref": "#/definitions/Employee",
                "required": true
              }
            }
          }
        },
        "description": "Makes a list of all employees",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "text/plain"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "text/plain": {}
            }
          },
          "400": {
            "description": "bad request",
            "content": {
              "text/plain": {}
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "text/plain": {}
            }
          }
        }
      }
    },
    "/update-employee/{id}": {
      "patch": {
        "summary": "Update specific fields of an employee by ID",
        "requestBody": {
          "description": "Employee data to be updated",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/EmployeePatchInput",
                "required": true
              }
            }
          }
        },
        "consumes": [
          "application/json"
        ],
        "produces": [
          "text/plain"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the employee",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "employee",
            "in": "body",
            "description": "Employee object containing specific fields to update",
            "schema": {
              "$ref": "#/definitions/EmployeePatchInput"
            }
          }
        ]
      },
      "responses": {
        "200": {
          "description": "Employee record updated successfully",
          "content": {
            "text/plain": {}
          },
          "400": {
            "description": "Invalid request payload",
            "content": {
              "text/plain": {}
            },
            "404": {
              "description": "Employee not found",
              "content": {
                "text/plain": {}
              },
              "500": {
                "description": "Internal server error",
                "content": {
                  "text/plain": {}
                }
              }
            }
          }
        }
      }
    },
    "/get-employee/{id}": {
      "get": {
        "summary": "Get an employee by ID",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Employee ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/Employee"
            }
          },
          "404": {
            "description": "Employee not found"
          }
        }
      }
    },
    "/delete-employee/{id}": {
      "delete": {
        "summary": "Delete an employee by ID",
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "description": "Employee ID",
          "required": true,
          "schema": {
            "type": "integer"
          }
        }
      ],
      "responses": {
        "204": {
          "description": "Employee deleted successfully"
        },
        "404": {
          "description": "Employee not found"
        }
      }
    }
    },
    "/login": {
      "post": {
        "summary": "Logs in a user",
        "requestBody": {
          "content": {
            "application/json": {
              "description": "lets in a user",
              "schema": {
                "$ref": "#/definitions/Employeelogin",
                "required": true
              }
            }
          }
        },
        "consumes": [
          "application/json"
        ],
        "produces": [
          "text/plain"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "text/plain": {}
            }
          },
          "400": {
            "description": "bad request",
            "content": {
              "text/plain": {}
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "text/plain": {}
            }
          }
        }
    }
  },
    "/create-employees": {
        "post": {
          "summary": "Create multiple employee records",
          "requestBody":{
          "content":{
            "application/json":{
              "description": "creating multiple employee records",
              "schema": {
                "$ref" : "#/definitions/Employees",
                "required": true
            }
            }
          }
        },
        "description": "Array of employee records",
          "consumes": ["application/json"],
          "produces": ["application/json"],
          "parameters": [],
          "responses": {
            "201": {
              "description": "Employee records created successfully"
            },
            "400": {
              "description": "Bad request"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
      "/refresh-token": {
        "post": {
          "summary": "Refresh JWT token",
          "description": "Refreshes an expired JWT token and provides a new one with a renewed expiration time.",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "refreshtoken": {
                      "type": "string",
                      "description": "The expired JWT token to be refreshed."
                    }
                  },
                  "required": ["token"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "token": {
                        "type": "string",
                        "description": "The new JWT token with a renewed expiration time."
                      }
                    }
                  }
                }
              }
            }
    }
 }
}
  },
 "definitions": {
  "Employee": {
    "type": "object",
    "properties": {
      "Department_id": {
        "type": "integer"
      },
      "first_name": {
        "type": "string"
      },
      "last_name": {
        "type": "string"
      },
      "email": {
        "type": "string"
      },
      "hire_date": {
        "type": "string"
      },
      "salary": {
        "type": "integer"
      },
      "password": {
        "type": "string"
      }
    }
  },
  "EmployeePatchInput": {
    "type": "object",
    "properties": {
      "salary": {
        "type": "integer"
      }
    }
  },
  "Employeelogin": {
    "type": "object",
    "properties": {
      "email": {
        "type": "string"
      },
      "password": {
        "type": "string"
      }
    }
  },
  "Employees" :{
    "type": "array",
    "items": {
      "type" :"object",
      "properties": {
        "Department_id": {
          "type": "integer"
        },
        "first_name": {
          "type": "string"
        },
        "last_name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "hire_date": {
          "type": "string"
        },
        "salary": {
          "type": "integer"
        },
        "password": {
          "type": "string"
        }
      },
      "required": ["Department_id","first_name","last_name","email","hire_date","salary","password"]
    }          
  }
}
    
    
}